\name{Combining attributes}
\alias{attribute.combination}
\concept{Vertex/edge/graph attributes}
\title{How igraph functions handle attributes when the graph changes}
\description{}
\details{}
\author{Gabor Csardi \email{csardi.gabor@gmail.com}}
\seealso{\code{\link[igraph]{attributes}} on how to use
  graph/vertex/edges attributes in general.}
\examples{}
\keyword{graphs}
