\name{ba.game}
\alias{ba.game}
\alias{barabasi.game}
\title{Generate scale-free graphs according to the Barabasi-Albert
  model}
\description{}
\usage{
ba.game(n, m = NULL, out.dist = NULL, out.seq = NULL, out.pref = FALSE, 
        \dots) 
}
\arguments{
  \item{n}{}
  \item{m}{}
  \item{out.dist}{}
  \item{out.seq}{}
  \item{out.pref}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
