\name{betweenness}
\alias{betweenness}
\alias{edge.betweenness}
\title{Vertex and edge betweenness centrality}
\description{}
\usage{
betweenness(graph, v = 1:vcount(graph), directed = TRUE)
edge.betweenness(graph, directed = TRUE)
}
\arguments{
  \item{graph}{}
  \item{v}{}
  \item{directed}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
