\name{closeness}
\alias{closeness}
\title{Closeness centrality of vertices}
\description{}
\usage{
closeness(graph, v = 1:vcount(graph), mode = "all")
}
\arguments{
  \item{graph}{}
  \item{v}{}
  \item{mode}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
