\name{community}
\alias{community}
\alias{cut.community}
\alias{eb.community}
\alias{eb.community2}
\alias{edge.type.matrix}
\alias{modularity}
\title{Undocumented functions for community structure detection}
\description{These functions are alpha code and also their syntax is
  expected to change. Don't use them!}
\usage{
eb.community(graph, directed = TRUE)
eb.community2(graph, directed = TRUE)
cut.community(graph, edges, after.removing)
edge.type.matrix(graph, types)
modularity(graph, types)
}
\arguments{
  \item{graph}{}
  \item{directed}{}
  \item{edges}{}
  \item{after.removing}{}
  \item{types}{}
}
% \details{}
\value{}  
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
% \seealso{}
\examples{}
\keyword{graphs}
