\name{components}
\alias{clusters}
\alias{cluster.distribution}
\alias{subcomponent}
\title{Connected components of a graph}
\description{}
\usage{
clusters(graph, mode="weak")
cluster.distribution(graph, cumulative = FALSE, mul.size = FALSE, \dots)
subcomponent(graph, v, mode = "all")
}
\arguments{
  \item{graph}{}
  \item{mode}{}
  \item{cumulative}{}
  \item{mul.size}{}
  \item{v}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
