\name{degree}
\alias{degree}
\alias{degree.distribution}
\title{Degree and degree distribution of the vertices}
\description{}
\usage{
degree(graph, v = 1:vcount(graph), mode = "total", loops = TRUE) 
degree.distribution(graph, cumulative = FALSE, \dots) 
}
\arguments{
  \item{graph}{}
  \item{v}{}
  \item{mode}{}
  \item{loops}{}
  \item{cumulative}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
