\name{degree.sequence.game}
\alias{degree.sequence.game}
\title{Generate random graphs with a given degree sequence}
\description{}
\usage{
degree.sequence.game(out.deg, in.deg = NULL, method = "simple", \dots) 
}
\arguments{
  \item{out.deg}{}
  \item{in.deg}{}
  \item{method}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
