\name{erdos.renyi.game}
\alias{erdos.renyi.game}
\alias{random.graph.game}
\title{Generate random graphs according to the Erdos-Renyi model}
\description{}
\usage{
erdos.renyi.game(n, p, directed = FALSE, loops = FALSE, \dots)
}
\arguments{
  \item{n}{}
  \item{p}{}
  \item{directed}{}
  \item{loops}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
