\name{plot.graph}
\alias{plot.graph}
\title{Plotting of graphs}
\description{}
\usage{
plot.graph(graph, layout = layout.random, layout.par = list(), 
    labels = NULL, label.color = "darkblue", label.font = NULL, 
    label.degree = -pi/4, label.dist = 0, vertex.color = "SkyBlue2", 
    vertex.size = 15, edge.color = "darkgrey", edge.width = 1, 
    axes = FALSE, xlab = "", ylab = "", \dots)
}
\arguments{
  \item{graph}{}
  \item{layout}{}
  \item{layout.par}{}
  \item{labels}{}
  \item{label.color}{}
  \item{label.font}{}
  \item{label.degree}{}
  \item{label.dist}{}
  \item{vertex.color}{}
  \item{vertex.size}{}
  \item{edge.color}{}
  \item{edge.width}{}
  \item{axes}{}
  \item{xlab}{}
  \item{ylab}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
