\name{print.graph}
\alias{print.graph}
\alias{summary.graph}
\title{Print graphs to the terminal}
\description{}
\usage{
\method{print}{graph}(graph)
\method{summary}{graph}(graph)
}
\arguments{
  \item{graph}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
