\name{read.graph}
\alias{read.graph}
\alias{read.graph.pajek}
\alias{read.graph.edgelist}
\title{Reading foreign file formats}
\description{}
\usage{
read.graph(file, format = "pajek", \dots)
}
\arguments{
  \item{file}{}
  \item{format}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
