\name{shortest.paths}
\alias{shortest.paths}
\alias{get.shortest.paths}
\title{Shortest (directed or undirected) paths between vertices}
\description{}
\usage{
shortest.paths(graph, v = 1:vcount(graph), mode = "all")
get.shortest.paths(graph, from = 1, mode = "all")
}
\arguments{
  \item{graph}{}
  \item{v}{}
  \item{mode}{}
  \item{from}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
