\name{simplify}
\alias{simplify}
\title{Remove loop and/or multiple edges from a graph}
\description{}
\usage{
simplify(graph, remove.loops = TRUE, remove.multiple = TRUE)
}
\arguments{
  \item{graph}{}
  \item{remove.loops}{}
  \item{remove.multiple}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
