\name{tkigraph}
\alias{tkigraph}
\concept{GUI}
\title{Experimental basic igraph GUI}
\description{This functions starts an experimental GUI to some igraph
  functions. The GUI was written in Tcl/Tk, so it is cross platform.}
\usage{
tkigraph()
}
%\arguments{}
\details{
  \code{tkigraph} has its own online help system, please see that for
  the details about how to use it.
}
\value{Returns \code{NULL}, invisibly.}
% \references{}
\author{ Gabor Csardi \email{Gabor.Csardi@unil.ch} }
\seealso{\code{\link{tkplot}} for interactive plotting of graphs.}
% \examples{}
\keyword{graphs}
