\name{community}
\alias{community}
\alias{community.cut}
\alias{community.eb}
\alias{community.eb2}
\alias{edge.type.matrix}
\alias{modularity}
\title{Undocumented functions for community structure detection}
\description{These functions are alpha code and also their syntax is
  expected to change. Don't use them!}
\usage{
community.eb(graph, directed = TRUE)
community.eb2(graph, directed = TRUE)
community.cut(graph, edges, after.removing)
edge.type.matrix(graph, types)
modularity(graph, types)
}
\arguments{
  \item{graph}{}
  \item{directed}{}
  \item{edges}{}
  \item{after.removing}{}
  \item{types}{}
}
% \details{}
\value{}  
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
% \seealso{}
\examples{}
\keyword{graphs}
