\name{conversion}
\alias{get.adjacency}
\alias{get.edgelist}
\title{Convert a graph to an adjacency matrix or an edge list}
\description{}
\usage{
get.adjacency(graph)
get.edgelist(graph)
}
\arguments{
  \item{graph}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
