\name{diameter}
\alias{diameter}
\title{Diameter of a graph}
\description{}
\usage{
diameter(graph, directed = TRUE, unconnected = TRUE)
}
\arguments{
  \item{graph}{}
  \item{directed}{}
  \item{unconnected}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
