\name{graph.structure}
\alias{add.edges}
\alias{add.vertices}
\alias{delete.edges}
\alias{delete.vertices}
\title{Method for structural manipulation of graphs}
\description{}
\usage{
add.edges(graph, edges)
add.vertices(graph, nv)
delete.edges(graph, edges)
delete.vertices(graph, v)
}
\arguments{
  \item{graph}{}
  \item{edges}{}
  \item{nv}{}
  \item{v}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
