\name{layout}
\alias{layout}
\alias{layout.random}
\alias{layout.circle}
\alias{layout.fruchterman.reingold}
\alias{layout.kamada.kawai}
\alias{layout.spring}
\title{Generate coordinates for plotting graphs}
\description{}
\usage{
layout.random(graph, params)
layout.circle(graph, params)
layout.fruchterman.reingold(graph, params)
layout.kamada.kawai(graph, params)
layout.spring(graph, params)
}
\arguments{
  \item{graph}{}
  \item{params}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}

