\name{local.scan}
\alias{local.scan}
\concept{Scan Statistics on Graphs}
\title{Compute local scan statistics on static and time-varying graphs}
\description{The scan statistic is a summary of the locality statistics
  that is computed from the local neighborhood of each vertex.} 
\usage{
local.scan (graph.us, graph.them = NULL, k = 1, FUN = NULL,
    weighted = FALSE, mode = c("out", "in", "all"),
    neighborhoods = NULL, ...) 
}
\arguments{
  \item{graph.us}{TODO}
  \item{graph.them}{TODO}
  \item{k}{The size of the local neighborhood for each vertex.}
  \item{FUN}{TODO}
  \item{weighted}{TODO}
  \item{mode}{TODO}
  \item{neighborhoods}{TODO}
  \item{\dots}{TODO}
}
\details{TODO}
\value{TODO}
\references{TODO}
\author{Gabor Csardi \email{csardi.gabor@gmail.com}}
\seealso{TODO}
\examples{TODO}
\keyword{graphs}
