\name{power.law.fit}
\alias{power.law.fit}
\title{Fitting a power-law distribution function to discrete data}
\description{}
\usage{
power.law.fit(x, xmin, start = 2, \dots)
}
\arguments{
  \item{x}{}
  \item{xmin}{}
  \item{start}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
