\name{structure.info}
\alias{vcount}
\alias{ecount}
\alias{neighbors}
\alias{is.directed}
\alias{are.connected}
\alias{igraph.type}
\title{Gaining information about graph structure}
\description{}
\usage{
vcount(graph)
ecount(graph)
neighbors(graph, v, mode = "out")
is.directed(graph)
are.connected(graph, v1, v2)
igraph.type(graph)
}
\arguments{
  \item{graph}{}
  \item{v}{}
  \item{mode}{}
  \item{v1}{}
  \item{v2}{}
}
\details{}
\value{}
\references{}
\author{Gabor Csardi \email{csardi@rmki.kfki.hu}}
\seealso{}
\examples{}
\keyword{graphs}
